feat(pricing-table): move functionality into pricing table (#8420)
